pyproject.toml
[project]
name = "beehivehub-python-sdk-runner"
version = "1.0.0"
description = "Interactive CLI tool for developers to validate and test the Beehive Hub Python SDK through manual API operations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "beehivehub-python-sdk @ git+https://github.com/paybeehive/beehivehub-python-sdk.git@first-version",
    "questionary>=2.1.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
beehive-runner = "src.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["src/__init__.py", "src/resources/__init__.py"]

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = [
    "if __name__ == .__main__.",
    "pragma: no cover",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
